migrated id's do bigint and RSU endpoint up to date

+++ database/mariadb/it2s_db.sql
@@ -1,44 +1,44 @@
 create database it2s_db;
 use databse it2s_db;
 create table it2s_db.Emitter(
-  station_id int not null,
+  station_id bigint not null,
   current_app_version int not null,
   primary key(station_id)
 );
 create table it2s_db.OBU(
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   last_power_status int not null,
   primary key(emitter_station_id),
   foreign key (emitter_station_id) references it2s_db.Emitter(station_id)
 );
 create table it2s_db.App(
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   configured_language char(2) not null,
   primary key(emitter_station_id),
   foreign key (emitter_station_id) references it2s_db.Emitter(station_id),
   check(configured_language in ('pt', 'en'))
 );
 create table it2s_db.Smartphone(
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   last_power_status int not null,
   primary key(emitter_station_id),
   foreign key (emitter_station_id) references it2s_db.App(emitter_station_id)
 );
 create table it2s_db.WebSite(
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   browser_version int not null,
   primary key(emitter_station_id),
   foreign key (emitter_station_id) references it2s_db.App(emitter_station_id)
 );
 create table it2s_db.RSU(--
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   latitude bigint not null,
   longitude bigint not null,
   primary key(emitter_station_id),
   foreign key (emitter_station_id) references it2s_db.Emitter(station_id)
 );
 create table it2s_db.CPM(--
-  rsu_station_id int not null,
+  rsu_station_id bigint not null,
   event_timestamp int not null,
   latitude bigint not null,
   longitude bigint not null,
@@ -51,7 +51,7 @@ create table it2s_db.CPM(--
   )
 );
 create table it2s_db.PerceivedObject(--
-  cpm_station_id int not null,
+  cpm_station_id bigint not null,
   event_timestamp int not null,
   perceived_object_id int not null,
   latitude bigint not null,
@@ -74,7 +74,7 @@ create table it2s_db.PerceivedObject(--
   )
 );
 create table it2s_db.CAM(--
-  station_id int not null,
+  station_id bigint not null,
   event_timestamp int not null,
   station_type int not null,
   speed int not null,
@@ -90,7 +90,7 @@ create table it2s_db.CAM(--
   )
 );
 create table it2s_db.VAM(--
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   event_timestamp int not null,
   station_type int not null,
   latitude bigint not null,
@@ -104,7 +104,7 @@ create table it2s_db.VAM(--
   )
 );
 create table it2s_db.DENM(--
-  emitter_station_id int not null,
+  emitter_station_id bigint not null,
   event_timestamp int not null,
   cause_code int not null,
   sub_cause_code int not null,
@@ -145,7 +145,7 @@ create table it2s_db.Notification1(
   )
 );
 create table it2s_db.Notification2(
-  cam_emitter_station_id int not null,
+  cam_emitter_station_id bigint not null,
   cam_event_timestamp int not null,
   status_id int not null,
   foreign key (cam_emitter_station_id, cam_event_timestamp) references it2s_db.CAM(station_id, event_timestamp),
